+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 2.4.4
+* `MouseHandler` と `TouchHandler` を `PointerPointerEvent` に統合するように
+* `InputHandlerLayer` に付加されていたクラス名 `input-handler` を削除
+* `InputHandlerLayer` に付加されていた tabindex の指定を削除
+
 ## 2.4.3
 * HTMLAudioPlugin で offset を利用できるように